% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_entry_to_tibble.R
\name{bib_entry_to_tibble}
\alias{bib_entry_to_tibble}
\title{Generate a tibble based on a .bib file}
\usage{
bib_entry_to_tibble(bib_file, entry = 1)
}
\arguments{
\item{bib_file}{Path to a `.bib` file.}

\item{entry}{Entry index or BibTeX key.}
}
\value{
a data frame with a standard set of fields for a single entry
}
\description{
Reads a BibTeX file and returns a data frame with a standard set of fields
for a single entry, either by numeric index
}
\examples{
\dontrun{
bib_entry_to_tibble("data/raw/sample_source.bib", 1)
}
}
