% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_extract_field.R
\name{bib_extract_field}
\alias{bib_extract_field}
\title{Extract a field of a BibTeX entry from a .bib file}
\usage{
bib_extract_field(entry_text, field)
}
\arguments{
\item{entry_text}{a value generate by bib_extract_field function}

\item{field}{the name of the field to extract the year from (default: "date").}
}
\value{
A integer containing the year
}
\description{
Reads a BibTeX file and returns a numeric of the year of publication
for a single entry, either by numeric index
}
\examples{
\dontrun{
entry_text = bib_read_entry("data/raw/sample_source.bib", 1)
bib_extract_year(entry_text)
}
}
